+++ CHANGELOG.md
@@ -35,6 +35,7 @@ This changelog keeps track of work items that have been completed and are ready 
 ### Improvements
 
 - **General**: TODO ([#TODO](https://github.com/kedacore/http-add-on/issues/TODO))
+- **Interceptor**: Add atomic endpoint readiness cache for O(1) hot-path checks ([#1472](https://github.com/kedacore/http-add-on/pull/1472))
 
 ### Fixes